requirements.txt
# Core Flask
Flask>=3.0.0
Flask-Login>=0.6.3
Werkzeug>=3.0.0

# WSGI Server for production
gunicorn>=21.2.0
waitress>=2.1.2

# Existing dependencies (keep all your current ones, add these if missing)
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
openpyxl>=3.1.0
python-dateutil>=2.8.2
pytz>=2023.3

# Email
secure-smtplib>=0.1.1

# Utilities
python-dotenv>=1.0.0

